connector: add connector type/logic
This connector sets the fundations for this library.

It needs access/secret keys, regions and eventually an aws.Config to be
created. If no aws.Config is provided, then a default one will be used.

Among the thing it does:
 * It understands 'complex' list of regions - with regex. Therefore 'eu-*'
and 'us-*' can both be passed to have all those regions available.
 * It saves the accountID for later use - such as AMI or Volume snapshots.
 * It instanciates sessions for later use of services: ELB, EC2, etc.

It should provide an abstraction easy enough to use all the services of
AWS in multiple regions without having to bother with their multiple
packages, connectors, etc.