main: add a simple main method as example
In order to use the library, only a couple parameters need to be set, in
fact the library needs as little as 1 call to be ready to be used.

A first one to get a connector object, and the other calls to directly
gather information from their AWS account/regions.