connector: use interface instead of struct
In order to mock AWS calls, we have to use interfaces, otherwise we
wouldn't be able to create our own methods to fake them.

The modification has also been followed by the update of:
svc.$SERVICE.ServiceName to $SERVICE.ServiceName

As the interface doesn't have this variable, and this variable is in
fact set as a const in the package, so there is no need to try to access
the variable from each service.