elasticache: update test naming and moved checkErrors
An extra test has been added to elasticache, as 'multiple regions with
error' was missing for the tags.

The formatting of test messages has been revised in order to comply
with: `$test-name [$index] - $element: received=%+v | expected=%+v`

```
-- FAIL: TestGetElasticacheTags (0.00s)
	elasticache_test.go:354: multiple region with error [3] - tags: received=[{

		} {
		  TagList: [{
		      Key: "test",
		      Value: "2"
		    }]
		}] | expected=[{

		} {
		  TagList: [{
		      Key: "test",
		      Value: "3"
		    }]
		}]
FAIL
```

The method checkErrors has also been moved to another file, as it will
be reused for other tests.