error: update error type and logic
The `RawsErr` did not make too much sense as we could simple defined a
`type Errs []Err`, thus simplifying the adding of errors as well as the
overall management/return.

The naming of errors in methods has been reviewed to be `errs`
everywhere, and those methods now check whether an error happened or
not before adding it (same for resources), in order to have a data more
representative of what really happens during calls.