s3: add methods to gather S3 buckets/objects/tags
These methods allow to get buckets/objects and their tags in multiple
regions. One improvement could be to provide a region for GetObjects
otherwise the Connector, will try to look it up through the regions.

That could be seen as both convenient and 'unecessary' as extra call are
done, but the client should know which regions they manage afterall.