test: add test for ec2, elb, elbv2, s3 and rds
Basic tests have been added to the various components currently used in
the project (ec2, elb, elbv2, s3 and rds). The mocking is fairly simple,
as the goal is to ensure that our methods are actually returning the
good set of data and/or errors.

Tests regarding accountID, and such things haven't been added as the
purpose is not to reproduce AWS behaviour - we trust them -, but simply
what our methods are doing.