billing: initial commit for importing billing data
Currently the backend used is DynamoDB, but it shouldn't be too
difficult to replace it in the future, the logic would remain the same.

Checker:
The checker is responsible for checking the hash of the S3 file and the
value in the dynamoDB. If those values are identical then no import is
needed.

Downloader:
The downloader is responsible for downloading and unzipping the billing
file is the different given paths.

Loader:
The loader reads the unzipped file and loads its data into a go struct,
in order to be consumed progressively via the 'injector'.

Injector:
The injector create the differents entries in dynamoDB for both the
report (in order to store the hash) and for the records (all fields are
being injected into the database), some extra one are also given such
as: a UUID or the name of the report the record comes from.

Manager:
The manager is the one calling the different component, it simply takes
the 2 AWS accounts (S3 & DynamoDB) as parameters, as well as the 'date'
and the 'bucket' it is supposed to import the data from.