error/log: add error handling and logs
The 'ProcessFile' file method now return errors whenever it faces them.
Those errors have different types so we can decide to log them, ignore
them or keep going.
The other methods from Checker/Downloader also return this type of error
now.

Logs have also been added in order to allow an easy reading, and the
capacity to set a loglevel, to filter information.

Finally some stats are being gathered during the processing of the file.
That could allow to create better entries for reports, so that we know
how many warnings/errors happened during the import, thus allowing us to
know if we should re-import it or not.