error/logs: improved log and errors management
The injector can now send request by pack of 25, and would return the
number of process items, as well as a list of unprocessed RecordId for
logging later on.

This created some changes in the 'saveRecords' / 'ProcessFile' method of
the Loader component as well as in the Manager. Those changes have
respectively be tested.