connector: Doc the errors of NewAWSReader
When using this service and wanted to actually know which errors can be
returned by the NewAWSReader, the implementation of it had to be
checked. For avoiding of having to do it each time by whoever needed it,
the errors that the NewAWSReader have been documented.

This PR has documented some internal functions because it was the way to
track from the bottom-up which errors the NewAWSReader can return, it
was not necessary for the end purpose, but having such docs are more
helpful than harmful.